requirements.txt
Django>=4.2.0
django-ckeditor>=6.0.0
Pillow>=9.0.0
openpyxl>=3.1.0
reportlab>=4.0.0
xlsxwriter>=3.1.0